+++ classes/elements.hpp
@@ -8,21 +8,21 @@
 class ElementContainer {
 private:
     Element** elements; // Массив указателей на элементы
-    size_t capacity;    // Максимальная вместимость контейнера
-    size_t size;        // Текущее количество элементов
+    int capacity;    // Максимальная вместимость контейнера
+    int size;        // Текущее количество элементов
 
     // Увеличение размера контейнера
     void resize();
 
 public:
     // Конструктор и деструктор
-    ElementContainer(size_t initialCapacity = 10);
+    ElementContainer(int initialCapacity = 10);
     ~ElementContainer();
 
     // Методы управления коллекцией
     void addElement(const Element& element);
-    void removeElement(size_t index);
-    void editElement(size_t index, const Element& updatedElement);
+    void removeElement(int index);
+    void editElement(int index, const Element& updatedElement);
     void findElementByMass(const int& mass) const;
 
     // Методы сохранения/загрузки
